Reuse buffers
In general objects created on functions (without new / malloc) are
stored in the stack, which for the ESP8266 is limited to 4k.

This change aims to decrease the size of ParamStream which is mostly
created on function calls, so on the stack (because it cannot be global
as it cannot be reused, recommended by JsonArduino).

https://github.com/esp8266/Arduino/issues/3597